[강운덕] [LUCYSUS-1744] spanId, parentSpanId를 int->long으로 다시 변경함.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-tomcat-profiler/trunk@2731 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/main/java/com/nhn/pinpoint/profiler/context/DefaultTrace.java
@@ -435,7 +435,7 @@ public final class DefaultTrace implements Trace {
     }
 
     @Override
-    public void recordNextSpanId(int nextSpanId) {
+    public void recordNextSpanId(long nextSpanId) {
         if (nextSpanId == -1) {
             return;
         }
